@@ -26,7 +26,7 @@
     "preferred-install": "dist"
   },
   "require": {
-    "php": ">=7.1.0",
+    "php": "^8.0",
     "ext-mbstring": "*",
     "ext-iconv": "*"
   },
